+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.7.6] - 2020-12-25
 ### Added
+- Soul speed enchantment token recipe
 - Brilliant trait to chaos gold
 ### Changed
 - Chaos rune and enchantment token recipes are now created with data gens (should have no changes)